feat: Add lazy loading to regular expressions (#16)
## Motivation

We should aim at limiting initial load time when importing the library,
one of the contributors to the initial load time are compilations of
regular expressions. We should avoid loading all of them at once, but
rather lazy load the ones which are used by users once.

## Related Changes

https://github.com/go-playground/validator/pull/1277

## Release Notes

All regular expressions are now lazy loaded, once when the corresponding
rule is used.