feat: Support validating slices of objects (#30)
## Motivation

Often times we want to validate a list of objects, not just one
particular instance.
While it is possible to manually range over a slice of structs and call
`govy.Validator.Validate` on each one of them, it's cumbersome and
should be solved be the library natively.

## Release Notes

Added new validator function for slices: `govy.Validator.ValidateSlice`.
It can be used to validate a slice of values, where each value has the
same type as the validator's type constraint. Internally it calls
`govy.Validator.Validate` on each element.