chore: Remove details from govy.RuleError (#60)
## Motivation

We're moving towards a more unified error message creation. The goal is
to allow users to manipulate the error message however they see fit.
With that in mind, `govy.RuleError` should have its `Message` field
populated with a created error that is ready to be used as is and passed
downstream.

## Related Changes

- https://github.com/nobl9/govy/pull/10
- https://github.com/nobl9/govy/pull/59

## Breaking Changes

Removed `govy.RuleError.Details` field, the error's details are now part
of the `govy.RuleError.Message` and not just the
`govy.RuleError.Error()` output.