chore: Simplify internal rules iteration withing PropertyRules (#80)
## Motivation

Previously the `steps` property of `PropertyRules` used to contain
different kinds of actions, like conditions, flow control and validation
rules.
When the logic was cleaned and only rules remained the field was left
intact as `[]interface{}`.
That is no longer necessary and we can safely convert it to a strictly
typed `[]validationInterface[T]`.