feat: Convert govy.ErrorCode to type definition (#85)
## Motivation

It would be easier to work with `ErrorCode` through dedicated methods
instead of plain functions which accept two error codes.

## Release Notes

`govy.ErrorCode` is now a type definition, it comes with two methods:
`Add`, which adds an error code to the chain and `Has`, which checks if
an error code is present in the chain.

## Breaking Changes

`govy.ErrorCode` has been converted from type alias to type definition.