feat: Add required as a synthetic rule in govy.Plan (#128)
## Motivation

Currently, If we mark a `govy.PropertyRules` as `Required()`, it won't
be reflected in the `govy.Plan` in any significant way.
Moreover, what If the same property is marked as required on a certain
predicate?

## Summary

In order to support the usage of predicates, the plan builder is adding
a synthetic `govy.Rule` with description set to "property is required"
and error code set to "required".

## Release Notes

Properties marked as `Required` will now be reported as having the
following rule in `govy.Plan`: `{"description":"property is
required","errorCode":"required"}`.