fix: Harden validation interface (#133)
## Motivation

Currently, due to how the `validationInterface` is used and defined, we
can potentially write this code:

```go
	v3 := govy.New(
		govy.For(func(f Foo) Foo { return f }).
			Rules(
				govy.For(func(f Foo) Foo { return f }),
			),
	)
```

This is obviously a wrong usage of the `govy.PropertyRules` ad the
`Rules()` method should only accept either `Rule` or `RuleSet`.

In order to fix this, we should define a dedicated interface for all 3
types of govy entities --> validator, properties and rules.

## Release Notes

`govy.PropertyRules.Rules()` and related functions will no longer accept
a `govy.Validator` instance.