fix: Correct govytest.Assert validator name matching (#144)
## Motivation

When the actual error checked by `govytest.AssertError` is of type
`govy.ValidatorError`, both validator name and index are ignored If we
set them in `govytest.ExpectedRuleError`.
This selective behavior is ambiguous and makes it harder check an
expected validator name is set for a given `govy.ValidatorError`.

## Release Notes

`govytest.AssertError` will now verify both validator name and index If
the actual error is of type `govy.ValidatorError` and if these values
are provided through `govytest.ExpectedRuleError`.